Updated skeleton for enemy abilities
