+++ Tower Defence/Assets/Scripts/Enemies/EnemyAbility.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Enemies
+{
+    public abstract class EnemyAbility : ScriptableObject
+    {
+        public abstract void Activate();
+    }
+}
