+++ Tower Defence/Assets/Scripts/Enemies/EnemyAbility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df653c58f4ef742bd842fc23ebe778ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
