+++ Tower Defence/Assets/Scripts/Enemies/EnemyWithAbilities.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Enemies
+{
+    public class EnemyWithAbilities : Enemy
+    {
+        public List<EnemyAbility> abilities;
+    }
+}
