+++ Tower Defence/Assets/Scripts/Enemies/HealAbility.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace Enemies
+{
+    [CreateAssetMenu(fileName = "HealRadius", menuName = "EnemyAbilities/Heal")]
+    public class HealAbility : EnemyAbility
+    {
+        public override void Activate()
+        {
+            
+        }
+    }
+}
