+++ Tower Defence/Assets/Scripts/Turrets/Upgrades/FasterShots.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Enemies;
 using UnityEngine;
 using UnityEngine.Serialization;
